Update README.md
Is Easter a good time to visit Egypt?
Tour Operator: Dav Safaris      Website: www.davsafaris.com       Email: info@davsafaris.com 
Tel: +256701412430 | +256757795781
Is Easter a good time to visit Egypt?
Is Easter a good time to visit Egypt? To fully embrace the Easter holiday safaris in Egypt is to rediscover the joy of childhood and walk the paths where Jesus once trod. Egypt, a perennial favorite among travelers, takes on a special charm during Easter, with the entire country adorned in festive decorations and exuding a spirit of joy and celebration. Egypt's Easter holiday safaris promises an unforgettable experience, allowing visitors to witness the magic and allure of its breathtaking natural and man-made wonders, each steeped in the ancient history of miracles.